requirements.txt
streamlit
transformers
torch
PyPDF2
bcrypt
python-dotenv
pytest
langchain
firebase-admin
tzlocal
pytest-mock
tf-keras
python-dotenv
PyMuPDF
openai==0.28

